2.1.0   * 動作環境を更新/Ruby 2.1.5, passenger 4.0.53   * スマートフォンでデフォルトサイズの画像を表示   * 複製更新記事の添付ファイルを差し替えても内容が変わらない不具合を修正   * FAQ検索結果画面でページネーションが発生した場合、エラーとなる不具合を修正   * TinyMCEのバージョンアップ。IE11に対応   * jQueryのバージョンアップ。IE9，10，11での表示不具合を解消   * 記事/新規作成で承認メールが送信されない不具合を修正   * 記事/コンテンツの名称を「自治体記事」に変更   * 記事/連絡先メールアドレス入力任意設定追加   * 記事/新着タブピースへ表示形式を追加   * 年間行事/表示順不具合を修正   * データ/テキスト・ファイル/全てのコンセプトから検索した場合に別サイトのデータが表示される不具合を修正   * デザイン/スタイルシート/設計者での操作不具合を修正   * サイト/ドメイン/新規登録時のindexページ作成不具合を修正   * 書き出しスクリプトの不具合を修正   * 読み上げ不具合を修正   * その他不具合を修正

+++ app/controllers/cms/admin/stylesheets_controller.rb
@@ -52,6 +52,7 @@ class Cms::Admin::StylesheetsController < Cms::Controller::Admin::Base
       params[:do] = "show"
       return show
     elsif request.post? && location = create
+      return error_auth unless @item.creatable?
       return redirect_to(location)
     end
     
@@ -67,11 +68,14 @@ class Cms::Admin::StylesheetsController < Cms::Controller::Admin::Base
   end
   
   def edit
+    return error_auth unless @item.editable?
+
     @item.read_body
     render :action => :edit
   end
   
   def create
+
     if params[:create_directory]
       if @item.create_directory(params[:item][:new_directory])
         flash[:notice] = 'ディレクトリを作成しました。'
@@ -92,6 +96,8 @@ class Cms::Admin::StylesheetsController < Cms::Controller::Admin::Base
   end
   
   def update
+    return error_auth unless @item.editable?
+
     old_path = @item.upload_path
     
     if @item.directory?
@@ -117,6 +123,8 @@ class Cms::Admin::StylesheetsController < Cms::Controller::Admin::Base
   end
   
   def move
+    return error_auth unless @item.editable?
+    
     if request.put?
       if @item.move(params[:item][:path])
         flash[:notice] = '移動処理が完了しました。'
@@ -129,6 +137,8 @@ class Cms::Admin::StylesheetsController < Cms::Controller::Admin::Base
   end
   
   def destroy
+    return error_auth unless @item.deletable?
+    
     if @item.destroy
       flash[:notice] = "削除処理が完了しました。"
     else
